* Some fixes...
* Forgot to check in package.json preference changes
* Handle flickering when first entering into subreddits
* Moved api calls back into the parents of PostList, PostList feels like it should be dumber.
* Adding keyboard shortcuts to fave and unfave.
