+++ extensions/reddit-search/src/Home.tsx
@@ -1,9 +1,15 @@
-import { ActionPanel, Action, Icon, List } from "@raycast/api";
-import { useState } from "react";
+import { ActionPanel, Action, Icon, List, showToast, Toast } from "@raycast/api";
+import { useEffect, useRef, useState } from "react";
 import { joinWithBaseUrl } from "./RedditApi/UrlBuilder";
 import SubredditList from "./SubredditList";
 import FilterBySubredditPostList from "./FilterBySubredditPostList";
 import PostList from "./PostList";
+import RedditResultItem from "./RedditApi/RedditResultItem";
+import RedditSort from "./RedditSort";
+import { AbortError } from "node-fetch";
+import getPreferences from "./Preferences";
+import { searchAll } from "./RedditApi/Api";
+import Sort from "./Sort";
 
 export default function Home({
   favorites,
@@ -14,12 +20,71 @@ export default function Home({
   addFavoriteSubreddit: (subreddit: string) => void;
   removeFavoriteSubreddit: (subreddit: string) => void;
 }) {
+  const abortControllerRef = useRef<AbortController | null>(null);
+  const [results, setResults] = useState<RedditResultItem[]>([]);
+  const [sort, setSort] = useState(RedditSort.relevance);
+  const [searchRedditUrl, setSearchRedditUrl] = useState("");
   const [searching, setSearching] = useState(false);
-  const [query, setQuery] = useState<string>("");
+  const queryRef = useRef<string>("");
+
+  const doSearch = async (query: string, sort = RedditSort.relevance, after = "") => {
+    abortControllerRef.current?.abort();
+    abortControllerRef.current = new AbortController();
+
+    setSearching(true);
+    queryRef.current = query;
+    setSort(sort);
+    if (!after) {
+      setResults([]);
+    }
+
+    if (!query) {
+      setSearching(false);
+      return;
+    }
+
+    try {
+      const preferences = getPreferences();
+      const apiResults = await searchAll(
+        "",
+        query,
+        preferences.resultLimit,
+        sort?.sortValue ?? "",
+        after,
+        abortControllerRef.current
+      );
+      setSearchRedditUrl(apiResults.url);
+
+      if (after) {
+        setResults([...results, ...apiResults.items]);
+      } else {
+        setResults(apiResults.items);
+      }
+    } catch (error) {
+      if (error instanceof AbortError) {
+        return;
+      }
+
+      console.log(error);
+      showToast({
+        style: Toast.Style.Failure,
+        title: "Something went wrong :(",
+        message: String(error),
+      });
+    } finally {
+      setSearching(false);
+    }
+  };
+
+  useEffect(() => {
+    return () => {
+      abortControllerRef?.current?.abort();
+    };
+  }, []);
 
   return (
-    <List isLoading={searching} onSearchTextChange={setQuery} throttle searchBarPlaceholder="Search Reddit...">
-      {!query && (
+    <List isLoading={searching} onSearchTextChange={doSearch} throttle searchBarPlaceholder="Search Reddit...">
+      {!queryRef.current && (
         <>
           <List.Section title="More ways to search">
             <List.Item
@@ -29,7 +94,7 @@ export default function Home({
               actions={
                 <ActionPanel>
                   <Action.Push
-                    title="Search subreddits"
+                    title="Search Subreddits"
                     target={
                       <SubredditList
                         favorites={favorites}
@@ -57,6 +122,7 @@ export default function Home({
                     <Action
                       title="Remove from Favorites"
                       icon={Icon.Trash}
+                      shortcut={{ modifiers: ["cmd"], key: "f" }}
                       onAction={async () => {
                         await removeFavoriteSubreddit(x);
                       }}
@@ -68,7 +134,13 @@ export default function Home({
           </List.Section>
         </>
       )}
-      <PostList setSearching={setSearching} query={query} />
+      <PostList
+        subreddit=""
+        posts={results}
+        sort={sort}
+        doSearch={(sort: Sort, after?: string) => doSearch(queryRef.current, sort, after)}
+        searchRedditUrl={searchRedditUrl}
+      />
     </List>
   );
 }
